requirements.txt
python-telegram-bot~=13.7
docker~=5.0.3